@@ -11,17 +11,17 @@ class MessageActionsModal extends StatefulWidget {
     required this.message,
     required this.messageWidget,
     required this.messageTheme,
-    this.showReactions = true,
-    this.showDeleteMessage = true,
-    this.showEditMessage = true,
+    this.showReactions,
+    this.showDeleteMessage,
+    this.showEditMessage,
     this.onReplyTap,
     this.onThreadReplyTap,
     this.showCopyMessage = true,
     this.showReplyMessage = true,
     this.showResendMessage = true,
-    this.showThreadReplyMessage = true,
-    this.showFlagButton = true,
-    this.showPinButton = true,
+    this.showThreadReplyMessage,
+    this.showFlagButton,
+    this.showPinButton,
     this.editMessageInputBuilder,
     this.reverse = false,
     this.customActions = const [],
@@ -47,34 +47,34 @@ class MessageActionsModal extends StatefulWidget {
   final MessageThemeData messageTheme;
 
   /// Flag for showing reactions
-  final bool showReactions;
+  final bool? showReactions;
 
   /// Callback when copy is tapped
   final OnMessageTap? onCopyTap;
 
   /// Callback when delete is tapped
-  final bool showDeleteMessage;
+  final bool? showDeleteMessage;
 
   /// Flag for showing copy action
   final bool showCopyMessage;
 
   /// Flag for showing edit action
-  final bool showEditMessage;
+  final bool? showEditMessage;
 
   /// Flag for showing resend action
   final bool showResendMessage;
 
   /// Flag for showing reply action
-  final bool showReplyMessage;
+  final bool? showReplyMessage;
 
   /// Flag for showing thread reply action
-  final bool showThreadReplyMessage;
+  final bool? showThreadReplyMessage;
 
   /// Flag for showing flag action
-  final bool showFlagButton;
+  final bool? showFlagButton;
 
   /// Flag for showing pin action
-  final bool showPinButton;
+  final bool? showPinButton;
 
   /// Flag for reversing message
   final bool reverse;
@@ -88,6 +88,8 @@ class MessageActionsModal extends StatefulWidget {
 
 class _MessageActionsModalState extends State<MessageActionsModal> {
   bool _showActions = true;
+  late List<String> _userPermissions;
+  late bool _isMyMessage;
 
   @override
   Widget build(BuildContext context) => _showMessageOptionsModal();
@@ -122,6 +124,19 @@ class _MessageActionsModalState extends State<MessageActionsModal> {
     final shiftFactor =
         numberOfReactions < 5 ? (5 - numberOfReactions) * 0.1 : 0.0;
 
+    final hasEditPermission = _userPermissions.contains(
+          PermissionType.updateAnyMessage,
+        ) ||
+        _userPermissions.contains(PermissionType.updateOwnMessage);
+
+    final hasDeletePermission = _userPermissions.contains(
+          PermissionType.deleteAnyMessage,
+        ) ||
+        _userPermissions.contains(PermissionType.deleteOwnMessage);
+
+    final hasReactionPermission =
+        _userPermissions.contains(PermissionType.sendReaction);
+
     final child = Center(
       child: SingleChildScrollView(
         child: Padding(
@@ -131,7 +146,7 @@ class _MessageActionsModalState extends State<MessageActionsModal> {
                 ? CrossAxisAlignment.end
                 : CrossAxisAlignment.start,
             children: <Widget>[
-              if (widget.showReactions &&
+              if ((widget.showReactions ?? hasReactionPermission) &&
                   (widget.message.status == MessageSendingStatus.sent))
                 Align(
                   alignment: Alignment(
@@ -168,21 +183,35 @@ class _MessageActionsModalState extends State<MessageActionsModal> {
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.stretch,
                       children: [
-                        if (widget.showReplyMessage &&
-                            widget.message.status == MessageSendingStatus.sent)
+                        if (widget.showReplyMessage ??
+                            (_userPermissions
+                                    .contains(PermissionType.quoteMessage) &&
+                                widget.message.status ==
+                                    MessageSendingStatus.sent))
                           _buildReplyButton(context),
-                        if (widget.showThreadReplyMessage &&
-                            (widget.message.status ==
-                                MessageSendingStatus.sent) &&
-                            widget.message.parentId == null)
+                        if (widget.showThreadReplyMessage ??
+                            _userPermissions
+                                    .contains(PermissionType.sendReply) &&
+                                (widget.message.status ==
+                                    MessageSendingStatus.sent) &&
+                                widget.message.parentId == null)
                           _buildThreadReplyButton(context),
                         if (widget.showResendMessage)
                           _buildResendMessage(context),
-                        if (widget.showEditMessage) _buildEditMessage(context),
+                        if (widget.showEditMessage ??
+                            _isMyMessage && hasEditPermission)
+                          _buildEditMessage(context),
                         if (widget.showCopyMessage) _buildCopyButton(context),
-                        if (widget.showFlagButton) _buildFlagButton(context),
-                        if (widget.showPinButton) _buildPinButton(context),
-                        if (widget.showDeleteMessage)
+                        if (widget.showFlagButton ??
+                            _userPermissions
+                                .contains(PermissionType.flagMessage))
+                          _buildFlagButton(context),
+                        if (widget.showPinButton ??
+                            _userPermissions
+                                .contains(PermissionType.pinMessage))
+                          _buildPinButton(context),
+                        if (widget.showDeleteMessage ??
+                            (_isMyMessage && hasDeletePermission))
                           _buildDeleteButton(context),
                         ...widget.customActions
                             .map((action) => _buildCustomAction(
@@ -603,7 +632,9 @@ class _MessageActionsModalState extends State<MessageActionsModal> {
                 widget.editMessageInputBuilder!(context, widget.message)
               else
                 MessageInput(
-                  editMessage: widget.message,
+                  messageInputController: MessageInputController(
+                    message: widget.message,
+                  ),
                   preMessageSending: (m) {
                     FocusScope.of(context).unfocus();
                     Navigator.pop(context);
@@ -642,5 +673,14 @@ class _MessageActionsModalState extends State<MessageActionsModal> {
         ),
       ),
     );
+  }
+
+  @override
+  void didChangeDependencies() {
+    final newStreamChannel = StreamChannel.of(context);
+    _userPermissions = newStreamChannel.channel.ownCapabilities;
+    _isMyMessage =
+        widget.message.user?.id == StreamChat.of(context).currentUser?.id;
+    super.didChangeDependencies();
   }
 }
