@@ -5,9 +5,9 @@ import 'package:rxdart/rxdart.dart';
 import 'package:stream_chat/src/client/retry_policy.dart';
 import 'package:stream_chat/stream_chat.dart';
 
-/// The retry queue associated to a channel
+/// The retry queue associated to a channel.
 class RetryQueue {
-  /// Instantiate a new RetryQueue object
+  /// Instantiate a new RetryQueue object.
   RetryQueue({
     required this.channel,
     this.logger,
@@ -17,13 +17,13 @@ class RetryQueue {
     _listenFailedEvents();
   }
 
-  /// The channel of this queue
+  /// The channel of this queue.
   final Channel channel;
 
-  /// The client associated with this [channel]
+  /// The client associated with this [channel].
   final StreamChatClient client;
 
-  /// The logger associated to this queue
+  /// The logger associated to this queue.
   final Logger? logger;
 
   late final RetryPolicy _retryPolicy;
@@ -63,7 +63,7 @@ class RetryQueue {
     }).addTo(_compositeSubscription);
   }
 
-  /// Add a list of messages
+  /// Add a list of messages.
   void add(List<Message> messages) {
     if (messages.isEmpty) return;
     if (!_messageQueue.containsAllMessage(messages)) {
@@ -113,6 +113,7 @@ class RetryQueue {
       } catch (e) {
         if (e is! StreamChatNetworkError || !e.isRetriable) {
           _messageQueue.removeMessage(message);
+          _sendFailedEvent(message);
           return true;
         }
         // retry logic
@@ -174,10 +175,10 @@ class RetryQueue {
     }
   }
 
-  /// Whether our [_messageQueue] has messages or not
+  /// Whether our [_messageQueue] has messages or not.
   bool get hasMessages => _messageQueue.isNotEmpty;
 
-  /// Call this method to dispose this object
+  /// Call this method to dispose this object.
   void dispose() {
     _messageQueue.clear();
     _compositeSubscription.dispose();
