@@ -481,14 +481,21 @@ void main() {
 
     final path = '${_getChannelUrl(channelId, channelType)}/truncate';
 
-    when(() => client.post(path)).thenAnswer(
-        (_) async => successResponse(path, data: <String, dynamic>{}));
+    when(() => client.post(
+              path,
+              data: {},
+            ))
+        .thenAnswer(
+            (_) async => successResponse(path, data: <String, dynamic>{}));
 
     final res = await channelApi.truncateChannel(channelId, channelType);
 
     expect(res, isNotNull);
 
-    verify(() => client.post(path)).called(1);
+    verify(() => client.post(
+          path,
+          data: {},
+        )).called(1);
     verifyNoMoreInteractions(client);
   });
 
