@@ -136,7 +136,9 @@ class ThreadPage extends StatelessWidget {
             ),
           ),
           MessageInput(
-            parentMessage: parent,
+            messageInputController: MessageInputController(
+              message: Message(parentId: parent!.id),
+            ),
           ),
         ],
       ),
