@@ -1,4 +1,3 @@
-import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:shimmer/shimmer.dart';
@@ -556,14 +555,8 @@ class _ChannelListViewState extends State<ChannelListView> {
                         );
                       },
               ),
-              if ([
-                'admin',
-                'owner',
-              ].contains(channel.state!.members
-                  .firstWhereOrNull(
-                    (m) => m.userId == channel.client.state.currentUser?.id,
-                  )
-                  ?.role))
+              if (channel.ownCapabilities
+                  .contains(PermissionType.deleteChannel))
                 IconSlideAction(
                   color: backgroundColor,
                   iconWidget: StreamSvgIcon.delete(
